add layout between main and advice
