+++ MyApplication3/app/src/main/AndroidManifest.xml
@@ -14,7 +14,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Voz4"></activity>
+        <activity android:name=".DisplayInfoActivity"></activity>
+        <activity android:name=".Voz4" />
         <activity android:name=".Voz3" />
         <activity
             android:name=".ui.login.Voz3"
@@ -46,7 +47,7 @@
                 android:value=".MainActivity" />
         </activity>
         <activity
-            android:name=".DisplayInfoActivity"
+            android:name=".DisplayAdviceActivity"
             android:configChanges="keyboardHidden|orientation|screenSize"
             android:parentActivityName=".MainActivity"
             android:screenOrientation="portrait"
