+++ MyApplication3/app/src/main/res/layout/activity_display_info.xml
@@ -6,65 +6,15 @@
     android:layout_height="match_parent"
     tools:context=".DisplayInfoActivity">
 
-    <ImageButton
-        android:id="@+id/imageButton"
-        android:layout_width="190dp"
-        android:layout_height="190dp"
-        android:layout_marginStart="32dp"
-        android:layout_marginTop="40dp"
-        android:layout_marginBottom="16dp"
-        android:onClick="fullImageVoz1"
-        android:scaleType="fitCenter"
-        app:layout_constraintBottom_toTopOf="@+id/imageButton3"
-        app:layout_constraintEnd_toStartOf="@+id/imageButton2"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.464"
-        app:srcCompat="@drawable/voz1_ico" />
-
-    <ImageButton
-        android:id="@+id/imageButton3"
-        android:layout_width="190dp"
-        android:layout_height="190dp"
-        android:layout_marginStart="32dp"
-        android:layout_marginTop="16dp"
-        android:layout_marginBottom="40dp"
-        android:onClick="fullImageVoz3"
-        android:scaleType="fitCenter"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/imageButton4"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/imageButton"
-        app:srcCompat="@drawable/voz3_ico" />
-
-    <ImageButton
-        android:id="@+id/imageButton4"
-        android:layout_width="190dp"
-        android:layout_height="190dp"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="32dp"
-        android:layout_marginBottom="40dp"
-        android:onClick="fullImageVoz4"
-        android:scaleType="fitCenter"
-        app:layout_constraintBottom_toBottomOf="parent"
+    <Button
+        android:id="@+id/button"
+        android:layout_width="232dp"
+        android:layout_height="110dp"
+        android:layout_marginTop="32dp"
+        android:onClick="openAdvice"
+        android:text="@string/open_advice"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/imageButton3"
-        app:layout_constraintTop_toBottomOf="@+id/imageButton2"
-        app:srcCompat="@drawable/voz4_ico" />
-
-    <ImageButton
-        android:id="@+id/imageButton2"
-        android:layout_width="190dp"
-        android:layout_height="190dp"
-        android:layout_marginTop="40dp"
-        android:layout_marginEnd="32dp"
-        android:layout_marginBottom="16dp"
-        android:onClick="fullImageVoz2"
-        android:scaleType="fitCenter"
-        app:layout_constraintBottom_toTopOf="@+id/imageButton4"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/imageButton"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/voz2_ico" />
-
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
