+++ MyApplication3/app/src/main/res/values/strings.xml
@@ -15,4 +15,5 @@
     <string name="invalid_username">Not a valid username</string>
     <string name="invalid_password">Password must be >5 characters</string>
     <string name="login_failed">"Login failed"</string>
+    <string name="open_advice">Открыть советы</string>
 </resources>
